Access block_state internals via `friend struct block_handle_accessor;`
